requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
dnspython==2.4.0
fastapi==0.100.0
h11==0.14.0
httpcore==0.17.3
idna==3.4
loguru==0.7.0
numpy==1.25.1
pinecone-client==2.2.2
pydantic==2.0.3
pydantic_core==2.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.1
win32-setctime==1.1.0
